signal_buffer: 64-bit cumulative head so burst_valid disambiguates ring laps (T44)
signal_buffer_burst_valid() compared ring-relative positions modulo total_cap
(~1M complex samples, ~420 ms). If the producer lapped the ring an extra whole
total_cap between a burst being tagged and the worker validating it, that extra
distance was invisible and a truly stale burst could alias back to an
apparently-fresh window — the worker would then decode overwritten samples.
dsp_processor.c also truncated the tagger's already-64-bit cumulative index to
uint32 before handing it on, discarding the lap information the check needed.

Fix: carry the tagger's full 64-bit cumulative index end to end and compare in
absolute space.
  - signal_buffer.c: add s_head_total, a 64-bit monotonic complex-sample
    counter advanced at the SAME single exit site as `head` (T2 index
    invariant holds for both). Published under a single-writer seqlock
    (__sync_synchronize fences, same idiom as usbring.c) so the Core-1 worker
    never sees a torn 64-bit read that could wrongly pass a stale burst.
  - signal_buffer_burst_valid() takes uint64_t start_idx and tests
    (head_total - start_idx) <= total_cap against the absolute window.
  - detected_burst_t.start_sample_idx widened to uint64_t; dsp_processor.c
    drops the (uint32_t) truncation.
  - worker_core1.c: check_start is 64-bit for burst_valid; ring-offset callers
    take (uint32_t)ext_start, congruent mod total_cap so the ring mapping is
    unchanged. Log formats widened to %llu.

For any tagger->worker latency below one lap this yields the identical verdict
to the old check, so it is behavior-neutral on the deterministic corpus and
only rejects the stale-beyond-one-lap case it previously mishandled.

Smoke-verified: RAW matched=62/65 exact=44 close=13 partial=4 div=1 BER=1.39% (baseline, no regression)

+++ p4-usb-host/main/dsp_processor.c
@@ -147,7 +147,7 @@ static void dispatch_gone_burst(dsp_processor_t *p, const fbt_burst_t *b)
     // (10·log10(mag² · HISTORY / baseline_sum) — see
     // fft_burst_tagger.c). Don't subtract noise_db.
     detected_burst_t out = {
-        .start_sample_idx = (uint32_t)b->start,
+        .start_sample_idx = b->start, // T44: keep full 64-bit cumulative index
         .length_samples   = length_u32,
         .rel_freq_hz      = rel_freq_hz,
         .peak_snr_db      = b->magnitude_db,
